Rename SearchInput files
