+++ src/core/Pagination/PageInput/PageInput.md
